Use regex from dev branch
The regex from the dev branch is more reliable, especially because ckeditor in D8 seems to adds ```<p>&nbsp;</p>```which is somehow not replaced on my page. Changing the regex to the one in the dev branch of emptyparagraphkiller does the trick.

Thanks for the port so far Flo, saved a lot of time!